Refactored google signing so it's done when going to calendar
